+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.4 (2013-05-14)
 
 - Updated the installation instructions and added **install.sh** as a post-download script.
+- Fixes to markdown template "hashes" now an attribute.
 
 ## 2.1.3 (2013-05-13)